+++ packages/whatsapp_bot_platform_interface/lib/src/model/export_models.dart
@@ -1,3 +1,4 @@
+export 'package:whatsapp_bot_platform_interface/src/model/whatsapp_event.dart';
 export 'package:whatsapp_bot_platform_interface/src/model/message_button.dart';
 export 'package:whatsapp_bot_platform_interface/src/model/call_event.dart';
 export 'package:whatsapp_bot_platform_interface/src/model/whatsapp_file_type.dart';