+++ packages/whatsapp_bot_platform_interface/lib/src/model/whatsapp_file_type.dart
@@ -1,1 +1,1 @@
-enum WhatsappFileType { document, image, audio }
+enum WhatsappFileType { document, image, audio, pdf, unknown }