+++ packages/whatsapp_bot_platform_interface/lib/src/wpp/wpp_events.dart
@@ -5,14 +5,6 @@ class WppEvents {
   WpClientInterface wpClient;
   WppEvents(this.wpClient);
 
-  // To get update of all messages
-  final StreamController<Message> messageEventStreamController =
-      StreamController.broadcast();
-
-  // To get update of all Calls
-  final StreamController<CallEvent> callEventStreamController =
-      StreamController.broadcast();
-
   // To get update of all Connections
   final StreamController<ConnectionEvent> connectionEventStreamController =
       StreamController.broadcast();
@@ -25,32 +17,9 @@ class WppEvents {
 
   void _onNewEvent(String eventName, dynamic eventData) {
     switch (eventName) {
-      case "messageEvent":
-        _onNewMessage(eventData);
-        break;
       case "connectionEvent":
         _onConnectionEvent(eventData);
         break;
-      case "callEvent":
-        _onCallEvent(eventData);
-    }
-  }
-
-  void _onNewMessage(msg) {
-    try {
-      Message message = Message.fromJson(msg);
-      messageEventStreamController.add(message);
-    } catch (e) {
-      WhatsappLogger.log("onMessageError : $e");
-    }
-  }
-
-  void _onCallEvent(call) {
-    try {
-      CallEvent callEvent = CallEvent.fromJson(call);
-      callEventStreamController.add(callEvent);
-    } catch (e) {
-      WhatsappLogger.log("onCallEvent : $e");
     }
   }