+++ packages/whatsapp_bot_flutter_mobile/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 2.0.0
+
+- update to whatsapp_bot_platform_interface 1.0.0
+- Breaking change: new way of listening to messages: client.on(WhatsappEvent.chat_new_message, (data) => print(data));
+- add support for listening to multiple events
+- fix delete message
+- add option to send file name in sendFileMessage
+- add more methods to WhatsappClient.chat
+
 ## 1.0.0
 
 - use whatsapp_bot_platform_interface