+++ packages/whatsapp_bot_platform_interface/CHANGELOG.md
@@ -1,3 +1,14 @@
+## 1.0.0
+
+* breaking changes: listen to new messages using client.on(WhatsappEvent.chat_new_message, (data){});
+* add option to send fileName with file
+* add more chat options
+* add better ways to subscribe to streams
+
+## 0.0.2
+
+* minor fixes
+
 ## 0.0.1
 
-* TODO: Describe initial release.
+* initial release.